# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Projects RBAC API
  description: |
    REST API for authentication, groups, membership, RBAC policies, and protected assets.

    Import this file into Postman (**Import → File**), Insomnia, Thunder Client, or any OpenAPI 3 tool.
  version: 1.0.0
  contact:
    name: Projects API
servers:
  - url: http://localhost:3000
    description: Local unified Next.js app (UI + API)
  - url: "{baseUrl}"
    description: Custom / cloud deployment
    variables:
      baseUrl:
        default: https://app.example.com
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorMessage:
      type: object
      properties:
        message:
          type: string
      required: [message]
    UserSummary:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
    RegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          example: alice
        email:
          type: string
          format: email
          example: alice@example.com
        password:
          type: string
          format: password
          example: Password123!
    LoginRequest:
      type: object
      required: [password]
      properties:
        username:
          type: string
          example: alice
        email:
          type: string
          format: email
          example: alice@example.com
        password:
          type: string
          format: password
          example: Password123!
    LoginResponse:
      type: object
      properties:
        message:
          type: string
        token:
          type: string
        sessionId:
          type: string
          format: uuid
        expiresAt:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/UserSummary'
    ApiError:
      type: object
      description: Standard API error response used by controllers and auth middleware.
      properties:
        message:
          type: string
        code:
          type: string
          enum:
            - NO_TOKEN
            - EXPIRED
            - REVOKED
            - INVALID
            - USER_NOT_FOUND
            - FORBIDDEN
            - INVALID_CREDENTIALS
            - CONFIG
            - VALIDATION
            - NOT_FOUND
            - CONFLICT
            - INTERNAL
        details:
          type: object
          additionalProperties: true
          description: Optional structured context (e.g. username, hint). Never required by clients.
      required: [message, code]
    AuthError:
      allOf:
        - $ref: '#/components/schemas/ApiError'
        - type: object
          properties:
            code:
              type: string
              enum: [NO_TOKEN, EXPIRED, REVOKED, INVALID, USER_NOT_FOUND, FORBIDDEN, INVALID_CREDENTIALS, PASSWORD_CHANGED, CONFIG]
    SessionRecord:
      type: object
      properties:
        _id:
          type: string
        sessionId:
          type: string
          format: uuid
        userId:
          type: string
        username:
          type: string
        userAgent:
          type: string
        ipAddress:
          type: string
        clientApp:
          type: string
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        lastSeenAt:
          type: string
          format: date-time
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          format: password
        newPassword:
          type: string
          format: password
          minLength: 8
    AdminChangePasswordRequest:
      type: object
      required: [newPassword]
      properties:
        newPassword:
          type: string
          format: password
          minLength: 8
    ResetPasswordRequest:
      type: object
      required: [newPassword]
      properties:
        newPassword:
          type: string
          format: password
    Group:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        description:
          type: string
        members:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateGroupRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: editors
        description:
          type: string
          example: Can edit content
...
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/UserSummary'
        '400':
          description: Validation or duplicate user error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login and create a DB-backed session JWT
      description: |
        Creates a MongoDB session document (shared auth store) and returns a JWT whose `sid`
        claim matches the session. Other apps can validate the same token via `/api/auth/validate`
        or by reading the `sessions` collection.
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Missing credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: Logout and revoke the current session
      operationId: logoutUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logged out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Current user, session, and effective permissions
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Authenticated profile
        '401':
          description: Session missing, expired, or revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
  /api/auth/validate:
    get:
      tags: [Auth]
      summary: Validate bearer session for shared authentication
      operationId: validateSession
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Session is valid
        '401':
          description: Session invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
  /api/auth/change-password:
    post:
      tags: [Auth]
      summary: Change password for the authenticated user
      description: Revokes all of the user's active sessions after a successful change.
      operationId: changeOwnPassword
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password updated
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
  /api/auth/sessions:
    get:
      tags: [Auth]
      summary: List active sessions
      description: |
        Users with `USER:READ` see all active sessions. Others see only their own.
      operationId: listSessions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Active sessions
          content:
            application/json:
              schema:
                type: object
                properties:
                  scope:
                    type: string
                    enum: [all, self]
                  sessions:
                    type: array
                    items:
                      $ref: '#/components/schemas/SessionRecord'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
  /api/auth/sessions/{sessionId}:
    delete:
      tags: [Auth]
      summary: Disconnect a session
      description: Users can disconnect their own sessions. `USER:WRITE` can disconnect any session.
      operationId: disconnectSession
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Session disconnected
        '403':
          description: Not allowed to disconnect this session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
...
  /api/auth/reset-password/{token}:
    post:
      tags: [Auth]
      summary: Reset password with token
      operationId: resetPassword
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '400':
          description: Invalid, expired, or missing password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
...
  /api/users/{id}/password:
    post:
      tags: [Users]
      summary: Admin reset of another user's password
      description: Requires `USER:WRITE`. Revokes all of the target user's sessions.
      operationId: adminChangeUserPassword
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminChangePasswordRequest'
      responses:
        '200':
          description: Password updated
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: User not found
  /api/groups:
    get:
      tags: [Groups]
      summary: List groups
      operationId: listGroups
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Group list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Group'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
    post:
      tags: [Groups]
      summary: Create group
      operationId: createGroup
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
      responses:
        '201':
          description: Group created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
  /api/groups/{id}:
    get:
      tags: [Groups]
      summary: Get group by ID
      operationId: getGroup
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Group found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '401':
          description: Unauthorized
        '404':
          description: Not found
    put:
      tags: [Groups]
      summary: Update group
      operationId: updateGroup
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
      responses:
        '200':
          description: Group updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '401':
          description: Unauthorized
        '404':
          description: Not found
    delete:
      tags: [Groups]
      summary: Delete group
      operationId: deleteGroup
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Group deleted
        '401':
          description: Unauthorized
        '404':
          description: Not found
...